Ensure that constant 1.0 is a float64 (and isn't converted to, for example, an int64)

+++ namesrv/named_test.go
@@ -52,7 +52,7 @@ func TestKillNamed(t *testing.T) {
 	const T = 1000
 	fn := sp.NAMED + "crashnd.sem"
 
-	e := crash.NewEventPath(crash.NAMED_CRASH, T, 1.0, fn)
+	e := crash.NewEventPath(crash.NAMED_CRASH, T, float64(1.0), fn)
 	err := crash.SetSigmaFail(crash.NewTeventMapOne(e))
 	assert.Nil(t, err)
